requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.2
django-filter       25.1
djangorestframework 3.16.0
Markdown            3.8
pillow              11.2.1
pip                 25.0
psycopg2-binary     2.9.10
sqlparse            0.5.3
